chore: renamed module to char
